python/mysql_connector/pyproject.toml
[tool.poetry]
name = "python_mysql_connector"
version = "0.1.0"
description = ""
authors = ["Weizhen Wang <wangweizhen@pingcap.com>"]

[tool.poetry.dependencies]
python = "^3.9"
mysql-connector = "^2.2.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
